+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -4,14 +4,15 @@
 
 ### Features Added
 
-- Added `QueryPartialErrorException` and `LogsQueryError` to handle errors.
-- Added `partial_error` and `is_error` attributes to `LogsQueryResult`.
-- Added an option `allow_partial_errors` that defaults to False, which can be set to not throw if there are any partial errors.
+- Added `LogsQueryPartialResult` and `LogsQueryError` to handle errors.
+- Added `partial_error` and `status` attributes to `LogsQueryResult`.
 - Added a new `LogsTableRow` type that represents a single row in a table.
 
 ### Breaking Changes
 
 - `LogsQueryResult` now iterates over the tables directly as a convinience.
+- `query` API now returns a union of `LogsQueryPartialResult` and `LogsQueryResult`.
+- `query_batch` API now returns a union of `LogsQueryPartialResult`, `LogsQueryError` and `LogsQueryResult`.
 - `metric_namespace` is renamed to `namespace` and is a keyword-only argument in `list_metric_definitions` API.
 
 ### Bugs Fixed